Moved loading-hint in application's source folder because it's not a library. Deleted the stringify library because it's not used. Removed the proxy definition form package.json
