+++ package.json
@@ -69,7 +69,6 @@
         "webpack-dev-server": "^3.7.2",
         "webpack-merge": "^4.2.1"
     },
-    "proxy": "http://localhost:7200",
     "dependencies": {
         "angular-animate": "1.3.8",
         "angular-bowser": "0.0.4",
