+++ src/js/angular/common/loading-hint.js
